added cast to long int to correct interger overflow

+++ astronauts/main.cpp
@@ -68,13 +68,13 @@ These totals are added and returned.
 
 */
 
-int comp(long total_astros, vector<vector<long>> pairs) {
+long int comp(int total_astros, vector<vector<int>> pairs) {
 
 	unordered_map<int, Group *> map; // key: astronaut #, value: Group object
 	unordered_set<Group *> root_group_set; // set of all groups which have no parent group
-	long num_of_known_astros = 0;
+	int num_of_known_astros = 0;
 	for (int i = 0; i < pairs.size(); i++) {
-		vector<long> astro(pairs[i]);
+		vector<int> astro(pairs[i]);
 
 		int map_size = map.size();
 		bool is_new_astro_1 = false;
@@ -143,25 +143,23 @@ int comp(long total_astros, vector<vector<long>> pairs) {
 			} //else if they are equal..ignore.
 		}
 	}
-	
-	long unknown_astro_size = total_astros - num_of_known_astros; // the number of astronauts where the group is unknown
-	long num_root_groups = root_group_set.size();
-	vector<long> root_group_size(num_root_groups);
-	long i = 0;
+
+	int unknown_astro_size = total_astros - num_of_known_astros; // the number of astronauts where the group is unknown
+	int num_root_groups = root_group_set.size();
+	vector<int> root_group_size(num_root_groups);
+	int i = 0;
 	for (auto itr = root_group_set.begin(); itr != root_group_set.end(); ++itr) {
 		root_group_size[i++] = ((*itr)->size);
 	}
 	
-	long num_of_combinations = 0;
+	long int num_of_combinations = 0;
 	for (i = 0; i < num_root_groups; i++) {
-		for (long j = i + 1; j < num_root_groups; j++) {
+		for (int j = i + 1; j < num_root_groups; j++) {
 			num_of_combinations += root_group_size[i] * root_group_size[j];
 		}
 		num_of_combinations += root_group_size[i] * unknown_astro_size;
 	}
-	num_of_combinations += unknown_astro_size * (unknown_astro_size - 1) / 2;
-
-
+    num_of_combinations += (long int) unknown_astro_size * (unknown_astro_size - 1) / 2;
 	return num_of_combinations;
 }
 
@@ -180,11 +178,11 @@ where:
 */
 
 int main() {
-	long n, p;
+	int n, p;
 	cin >> n;
 	cin >> p;
-	vector<vector<long>> v;
-	long x, y;
+	vector<vector<int>> v;
+	int x, y;
 	for (int i = 0; i < p; i++) {
 		x = 0; y = 0;
 		cin >> x;
